+++ resolver.go
@@ -129,7 +129,11 @@ func (r *Resolver) resolveContentManagerWrapper() *ContentManagerWrapper {
 		return r.contentManagerWrapper
 	}
 
-	r.contentManagerWrapper = NewContentManagerWrapper(r.ResolveStore(), r.resolveCarManager(), r.resolveTrackManager())
+	r.contentManagerWrapper = NewContentManagerWrapper(
+		r.ResolveStore(),
+		r.resolveCarManager(),
+		r.resolveTrackManager(),
+	)
 
 	return r.contentManagerWrapper
 }