+++ store.go
@@ -61,3 +61,19 @@ type Store interface {
 	UpsertKissMyRankOptions(kmr *KissMyRankConfig) error
 	LoadKissMyRankOptions() (*KissMyRankConfig, error)
 }
+
+func loadChampionshipRaceWeekends(championship *Championship, store Store) error {
+	var err error
+
+	for _, event := range championship.Events {
+		if event.IsRaceWeekend() {
+			event.RaceWeekend, err = store.LoadRaceWeekend(event.RaceWeekendID.String())
+
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
+}